shave 4 bytes off the GelETree (at least on x86) The savings will be
Fri Sep 26 18:46:32 2003  George Lebl <jirka@5z.com>

	* src/structs.h, src/eval.c: shave 4 bytes off the GelETree (at least
	  on x86) The savings will be invisible on archs with alignment.  So
	  we're down at 20 bytes per GelETree node (plus any data of course)

+++ src/eval.c
@@ -5569,8 +5569,8 @@ gather_comparisons(GelETree *n)
 				ali->any.next = NULL;
 				nn->comp.nargs++;
 				nn->comp.comp =
-					g_slist_append(nn->comp.comp,
-						      GINT_TO_POINTER(n->op.oper));
+					g_slist_append (nn->comp.comp,
+							GINT_TO_POINTER((int)n->op.oper));
 
 				freenode(n);
 				n = t;
